Remove all @codingStandardsIgnore… exclusions
It looks like these exclusions have been placed in 2015 in the
original patch that added PHPCS to this codebase. It looks
like this was done because of the underscores in the property
and method names. Unfortunately this disables many, many more
sniffs.

In this first patch of a series of cleanup patches I do
nothing but remove all inline exclusions, and (temporarily)
replace them with exclusions in the .phpcs.xml file.

The only other change is the removal of underscores from
method names. This was a convention to mark methods as
private, back in the day when there was no "private" keyword
in PHP.

You might wonder if the callback functions referenced in
Farmer.php need to be public. No. The callers (look for
call_user_func) are all in the same file. Hence private.

Change-Id: I285d5033d0fe0494f499ecc2c53cf3cdde63beb7